Remove int enforcement in middle step of buy/sell price calculation to prevent rounding errors

See merge request rainlash/lt-maker!443

+++ app/engine/component_system_compiler/item_system_base.py
@@ -7,11 +7,11 @@ class Defaults():
         return None
 
     @staticmethod
-    def buy_price(unit, item) -> int:
+    def buy_price(unit, item) -> float:
         return None
 
     @staticmethod
-    def sell_price(unit, item) -> int:
+    def sell_price(unit, item) -> float:
         return None
 
     @staticmethod
